Nullop refactor to make the exporter module (previously gif render module) more generic, to facilitate uploading to server later.
